🐛 fix: fix typings
