+++ src/constraints/symbolParams/Modal.ts
