+++ src/constraints/symbolParams/WelcomeHeader.ts
