+++ src/constraints/symbolParams/index.ts
