+++ src/types/index.ts
@@ -1,5 +1,6 @@
-export { SketchFormat } from './sketchFormat';
 export * from './type';
 export * from './model';
 export * from './layout';
 export * from './symbol';
+export * from './gradient';
+export { SketchFormat } from './sketchFormat';
